requirements.txt
torch>=2.0
torchvision>=0.15
numpy
pandas
matplotlib
seaborn
scikit-learn
Pillow
pyyaml
tqdm
